Split WorkflowExecutor to WorkflowExectorMain
