+++ digdag-core/build.gradle
@@ -49,4 +49,5 @@ dependencies {
     compile 'org.apache.httpcomponents:httpclient:4.5.5' // depends on aws-java-sdk 1.11.545
 
     testCompile project(path: ':digdag-client', configuration: 'testArtifacts')
+    testCompile project(path: ':digdag-svc-executor', configuration: 'testArtifacts')
 }
