+++ digdag-svc-executor/build.gradle
@@ -2,6 +2,7 @@
 dependencies {
     compile project(':digdag-spi')
     compile project(':digdag-client')
+    compile project(':digdag-core')
     compile project(':digdag-plugin-utils')
     compile project(':digdag-guice-rs-server')
